refactor: consolidate email dispatcher handler param types (PIN-10067)
Move the per-event handler param aliases (AgreementHandlerParams,
PurposeHandlerParams, etc.) from src/handlers/handlerCommons.ts into
src/models/handlerParams.ts so that all handler param types live in a
single file alongside the generic HandlerCommonParams / HandlerParams<T>
aliases. Delete the now-empty handlerCommons.ts and update the handler
imports accordingly.
