+++ packages/email-notification-dispatcher/src/handlers/agreements/handleAgreementUnsuspendedByPlatformToConsumer.ts
@@ -13,7 +13,7 @@ import {
   mapRecipientToEmailPayload,
   retrieveAgreementEservice,
 } from "pagopa-interop-notification-commons";
-import { AgreementHandlerParams } from "../handlerCommons.js";
+import { AgreementHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
