+++ packages/email-notification-dispatcher/src/handlers/authorization/handleClientKeyAdded.ts
@@ -15,7 +15,7 @@ import {
   mapRecipientToEmailPayload,
   clientKeyNotFound,
 } from "pagopa-interop-notification-commons";
-import { ClientKeyHandlerParams } from "../handlerCommons.js";
+import { ClientKeyHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
