+++ packages/email-notification-dispatcher/src/handlers/authorization/handleClientPurposeRemovedEvent.ts
@@ -12,7 +12,7 @@ import {
   retrievePurpose,
   mapRecipientToEmailPayload,
 } from "pagopa-interop-notification-commons";
-import { ClientPurposeHandlerParams } from "../handlerCommons.js";
+import { ClientPurposeHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
