+++ packages/email-notification-dispatcher/src/handlers/authorization/handleProducerKeychainEserviceAdded.ts
@@ -13,7 +13,7 @@ import {
   mapRecipientToEmailPayload,
   eserviceNotFound,
 } from "pagopa-interop-notification-commons";
-import { ProducerKeychainEServiceHandlerParams } from "../handlerCommons.js";
+import { ProducerKeychainEServiceHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
