+++ packages/email-notification-dispatcher/src/handlers/authorization/handleProducerKeychainKeyAdded.ts
@@ -13,7 +13,7 @@ import {
   mapRecipientToEmailPayload,
   producerKeychainKeyNotFound,
 } from "pagopa-interop-notification-commons";
-import { ProducerKeychainKeyHandlerParams } from "../handlerCommons.js";
+import { ProducerKeychainKeyHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
