+++ packages/email-notification-dispatcher/src/handlers/eserviceTemplates/handleEserviceTemplateVersionPublished.ts
@@ -16,7 +16,7 @@ import {
   getRecipientsForTenants,
   mapRecipientToEmailPayload,
 } from "pagopa-interop-notification-commons";
-import { EserviceTemplateHandlerParams } from "../handlerCommons.js";
+import { EserviceTemplateHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
