+++ packages/email-notification-dispatcher/src/handlers/eserviceTemplates/handleEserviceTemplateVersionSuspendedToCreator.ts
@@ -13,7 +13,7 @@ import {
   getRecipientsForTenants,
   mapRecipientToEmailPayload,
 } from "pagopa-interop-notification-commons";
-import { EserviceTemplateHandlerParams } from "../handlerCommons.js";
+import { EserviceTemplateHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
