+++ packages/email-notification-dispatcher/src/handlers/purposes/handlePurposeVersionRejectedOtherVersion.ts
@@ -13,7 +13,7 @@ import {
   getRecipientsForTenants,
   mapRecipientToEmailPayload,
 } from "pagopa-interop-notification-commons";
-import { PurposeHandlerParams } from "../handlerCommons.js";
+import { PurposeHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
