+++ packages/email-notification-dispatcher/src/handlers/tenants/handleTenantVerifiedAttributeRevoked.ts
@@ -15,7 +15,7 @@ import {
   mapRecipientToEmailPayload,
   retrieveAttribute,
 } from "pagopa-interop-notification-commons";
-import { TenantHandlerParams } from "../handlerCommons.js";
+import { TenantHandlerParams } from "../../models/handlerParams.js";
 
 import { config } from "../../config/config.js";
